Refactor VulkanBuffer
